# frontend/public/openapi.yaml
openapi: 3.0.0
info:
  title: DevPulse API
  description: |
    API for the DevPulse platform covering authentication, repository analysis, and AI Copilot integrations.

    ### Security Notes
    All authenticated endpoints require a valid JWT bearer token.
    Tokens are validated server-side and expire after a fixed duration.

    ### Implementation & Safety Notes (Internal & Auditor References)
    * **CORS Restrictions**: The backend strictly limits incoming CORS requests to approved origins (such as `https://devpulse.dev` and local development environments) and matches subdomains securely.
    * **Token Validation**: Authentication is strictly verified server-side on every request using a stateless JWT validator; database-backed blacklists or custom checks ensure immediate revocation.
    * **Rate Limiting**: To prevent abuse and Denial of Service (DoS), endpoints are bound by sliding-window rate limiting (100 requests/minute default, with specialized limits like `/v1/analyze` restricted to 5 requests/24h per user).
    * **Input Sanitization**: User-supplied input variables (such as `repositoryFullName` and `repoUrl` in analysis runs) are thoroughly sanitized and strictly validated against regex patterns to eliminate path traversal, SSRF, and command injection vectors before entering the analysis pipeline.
  version: 1.0.0
  x-rateLimit:
    limit: 100 requests/minute
  contact:
    name: DevPulse Security Team
    email: security@devpulse.dev
  license:
    name: MIT

servers:
  - url: https://api.devpulse.dev
    description: Production API

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Unauthorized access'
        details:
          type: string
          example: 'Token expired'

  responses:
    UnauthorizedError:
      description: Access token is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Access forbidden. Insufficient permissions or scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded (100 requests/minute).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: An unexpected internal server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

security:
  - bearerAuth: []

paths:
  /v1/auth/github:
    get:
      summary: Initiate GitHub OAuth
      description: Redirects the user to the GitHub authorization page.
      security: []
      responses:
        '302':
          description: Redirect to GitHub.

  /v1/auth/github/callback:
    get:
      summary: GitHub OAuth Callback
      description: Handles the callback from GitHub, exchanges code for token, and redirects back to the frontend after secure authentication flow.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirects back to the frontend after secure authentication flow.

  /v1/auth/me:
    get:
      summary: Get Current User
      description: Returns the authenticated user's profile and GitHub token sync status.
      responses:
        '200':
          description: User info retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  githubTokenSynced:
                    type: boolean
                  user:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/repos:
    get:
      summary: List Repositories
      description: Fetches a list of the user's GitHub repositories.
      responses:
        '200':
          description: List of repositories.
          content:
            application/json:
              schema:
                type: object
                properties:
                  repositories:
                    type: array
                    items:
                      type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/analyze:
    post:
      summary: Analyze Repository
      description: Runs the fast AI repository metadata analysis for a specified repository.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - repositoryFullName
              properties:
                repositoryFullName:
                  type: string
                  example: SSSahil15/DevPulse
                repoUrl:
                  type: string
                  example: https://github.com/SSSahil15/DevPulse
      responses:
        '200':
          description: Analysis returned.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/pipeline/scan:
    post:
      summary: Simulate CI/CD Scan
      description: Creates an async scan job that runs Trivy, calculates a DevPulse score, and stores a pipeline result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - repositoryFullName
              properties:
                repositoryFullName:
                  type: string
                  example: SSSahil15/DevPulse
                repoUrl:
                  type: string
                  example: https://github.com/SSSahil15/DevPulse
      responses:
        '202':
          description: Scan job accepted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/pipeline/jobs/{jobId}:
    get:
      summary: Get Simulation Status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            example: job_9a2f1c8e0b4d2a6f
      responses:
        '200':
          description: Scan job status.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/pipeline/results:
    get:
      summary: List Pipeline Results
      description: Returns historical analysis results.
      responses:
        '200':
          description: A list of pipeline results.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/ai/chat:
    post:
      summary: Ask AI Copilot
      description: Submit a natural language query to the AI Copilot regarding codebase health.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  example: 'How do I fix the SQL injection vulnerability?'
                context:
                  type: object
                history:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: AI response returned.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'